Changed matchbox recipe and display from int to double

+++ src/tuner.h
@@ -11,12 +11,12 @@ public:
     Tuner();
 
     int getActualPosition() const;
-    void setActualPosition(int position);
+    void setActualPosition(double position);
 
     int getRecipePosition() const;
     QString getRecipePositionQStr() const;
-    void setRecipePosition(int setpoint);// sends value to the controller
-    void updateRecipePosition(int setpoint);// updates the ui recipe value
+    void setRecipePosition(double setpoint);// sends value to the controller
+    void updateRecipePosition(double setpoint);// updates the ui recipe value
 
     double getDefaultRecipe() const;
     void setDefaultRecipe(double setpoint);
@@ -27,16 +27,16 @@ public:
     void updateAutoTune(bool state);
 
 signals:
-    void recipePositionChanged(const int recipePosition);// sends value to the controller
-    void updateUIRecipePosition(const int m_recipePosition);// updates the ui recipe value
-    void actualPositionChanged(const int actualPosition);
+    void recipePositionChanged(const double recipePosition);// sends value to the controller
+    void updateUIRecipePosition(const double m_recipePosition);// updates the ui recipe value
+    void actualPositionChanged(const double actualPosition);
     void defaultRecipeChanged(const double defaultRecipe);
     void autoTuneChanged(const bool autoTune);
     void updateUIAutoTune(const bool m_autoTune);
 
 private:
-    int m_actualPosition;
-    int m_recipePosition;
+    double m_actualPosition;
+    double m_recipePosition;
     double m_defaultRecipe;
     bool m_autoTune;
 };
